# add @adjoin for relu to increase speed