--- a/src/lib/nnlib.jl
+++ b/src/lib/nnlib.jl
@@ -1,5 +1,18 @@
 using NNlib
-import NNlib: softmax, ∇softmax, logsoftmax, ∇logsoftmax, conv, depthwiseconv, ∇conv_data, ∇depthwiseconv_data, maxpool, meanpool, σ
+import NNlib: softmax, ∇softmax, logsoftmax, ∇logsoftmax, conv, depthwiseconv, ∇conv_data, ∇depthwiseconv_data, maxpool, meanpool, σ, relu
+
+Zygote.@adjoint function Base.broadcasted(::typeof(relu), x::Array{T}) where T<:Real
+    y = relu.(x)
+    return y, Δ -> begin
+        res = zero(Δ)
+        @inbounds for i in 1:length(res)
+            if y[i] > 0
+                res[i] = Δ[i]
+            end
+        end
+        (nothing, res)
+    end
+end
 
 @adjoint function σ(x::Real)
     y = σ(x)