--- a/src/lib/nnlib.jl
+++ b/src/lib/nnlib.jl
@@ -1,9 +1,9 @@
 using NNlib
 import NNlib: softmax, ∇softmax, logsoftmax, ∇logsoftmax, conv, depthwiseconv, ∇conv_data, ∇depthwiseconv_data, maxpool, meanpool, σ, relu
 
-@adjoint function Base.Broadcast.broadcasted(::typeof(relu), xs::Numeric{T}) where T<:Number
-    ys = relu.(xs)
-    ys, Δ -> (nothing, broadcast((y, d) -> y > 0 ? d : zero(T), ys, Δ))
+@adjoint function Base.Broadcast.broadcasted(::typeof(relu), x::Numeric{T}) where T<:Number
+    y = relu.(x)
+    y, Δ -> (nothing, broadcast((z, d) -> z > 0 ? d : zero(T), y, Δ))
 end
 
 @adjoint function σ(x::Real)