--- a/src/lib/nnlib.jl
+++ b/src/lib/nnlib.jl
@@ -1,6 +1,8 @@
 using NNlib
 import NNlib: softmax, ∇softmax, logsoftmax, ∇logsoftmax, conv, depthwiseconv, ∇conv_data, ∇depthwiseconv_data, maxpool, meanpool, σ, relu
 
+@adjoint Base.Broadcast.broadcasted(::typeof(identity), x::Numeric{T}) where T<:Number = x, Δ -> (nothing, Δ)
+
 @adjoint function Base.Broadcast.broadcasted(::typeof(relu), x::Numeric{T}) where T<:Number
     y = relu.(x)
     y, Δ -> (nothing, broadcast((z, d) -> z > 0 ? d : zero(T), y, Δ))