--- a/src/lib/nnlib.jl
+++ b/src/lib/nnlib.jl
@@ -1,8 +1,6 @@
 using NNlib
 import NNlib: softmax, ∇softmax, logsoftmax, ∇logsoftmax, conv, depthwiseconv, ∇conv_data, ∇depthwiseconv_data, maxpool, meanpool, σ, relu
 
-@adjoint Base.Broadcast.broadcasted(::typeof(identity), x::Numeric) = x, Δ -> (nothing, Δ)
-
 @adjoint function Base.Broadcast.broadcasted(::typeof(relu), x::Numeric)
     y = relu.(x)
     y, Δ -> (nothing, ifelse.(y .> 0, Δ, zero.(y)))