--- a/test/gradcheck.jl
+++ b/test/gradcheck.jl
@@ -29,6 +29,9 @@ Random.seed!(0)
 
 @test gradient(//, 2, 3) === (1//3, -2//9)
 
+@test gradtest((x, W, b) -> identity.(W*x .+ b), 5, (2,5), 2)
+@test gradtest((x, W, b) -> identity.(W*x .+ b), (5,3), (2,5), 2)
+
 @test gradtest((x, W, b) -> relu.(W*x .+ b), 5, (2,5), 2)
 @test gradtest((x, W, b) -> relu.(W*x .+ b), (5,3), (2,5), 2)