```julia

using Flux, Zygote, BenchmarkTools

x = rand(128)

# before
@benchmark gradient((x) -> sum(relu.(x)), $x)

#=
BenchmarkTools.Trial: 
  memory estimate:  15.41 KiB
  allocs estimate:  425
  --------------
  minimum time:     9.980 μs (0.00% GC)
  median time:      10.649 μs (0.00% GC)
  mean time:        12.495 μs (4.12% GC)
  maximum time:     2.701 ms (98.24% GC)
  --------------
  samples:          10000
  evals/sample:     1
=#

# after

@benchmark gradient((x) -> sum(relu.(x)), $x)
#=
BenchmarkTools.Trial: 
  memory estimate:  2.30 KiB
  allocs estimate:  3
  --------------
  minimum time:     258.013 ns (0.00% GC)
  median time:      303.099 ns (0.00% GC)
  mean time:        368.242 ns (16.51% GC)
  maximum time:     4.796 μs (85.99% GC)
  --------------
  samples:          10000
  evals/sample:     394
=#
```