@MikeInnes 2019-10-09
This is useful, but you probably don't need the loop vs. just using another broadcast in the backwards pass. See also the gradients in `broadcast.jl` for examples.