@MikeInnes 2019-10-24
Thanks @findmyway!

@pevnak yeah, those would be great to have as a PR.

bors r+